Various bugfixes. A couple improvements.
Fix an issue where tuples could be constructed with items in reverse order.

Fix an issue where the offset into Torch storage didn't reflect the element size.

Add the ability to get the element size to TensorType.

Implement IntoIterator for RepugnantTorchTensors.

Fix a typo in the resolver definition in Cargo.toml.